+++ src/api.py
@@ -7,7 +7,7 @@ import inspect
 import json
 import time
 import traceback
-from datetime import datetime
+from datetime import datetime, timezone
 
 import psutil
 from fastapi import Request
@@ -128,14 +128,14 @@ async def tracebackHandler(request: Request, exc: Exception, err: str):
             if app.redis.lpos("session_errs", err_hash) is None:
                 app.redis.lpush("session_errs", err_hash)
 
-            logger.error(f"[{app.config.abbr}] {err_hash} [DATABASE] [{str(datetime.now())}]\nRequest IP: {request.client.host}\nRequest URL: {str(request.url)}\n{err}")
+            logger.error(f"[{app.config.abbr}] {err_hash} [DATABASE] [{datetime.now(timezone.utc).isoformat()}]\nRequest IP: {request.client.host}\nRequest URL: {str(request.url)}\n{err}")
 
             if int(time.time()) - app.db.POOL_START_TIME >= 60 and app.db.POOL_START_TIME != 0:
                 app.state.dberr.append(time.time())
                 app.state.dberr[:] = [i for i in app.state.dberr if i > time.time() - 1800]
 
                 if len(app.state.dberr) % 50 == 0:
-                    opqueue.queue(app, "post", app.config.webhook_error, app.config.webhook_error, json.dumps({"embeds": [{"title": "Database Error", "description": "Detected too many database errors. It's recommended to restart service.", "fields": [{"name": "Host", "value": app.config.domain, "inline": True}, {"name": "Unique ID", "value": app.config.abbr, "inline": True}, {"name": "Version", "value": app.version, "inline": True}], "color": int(app.config.hex_color, 16), "timestamp": str(datetime.now())}]}), {"Content-Type": "application/json"}, None)
+                    opqueue.queue(app, "post", app.config.webhook_error, app.config.webhook_error, json.dumps({"embeds": [{"title": "Database Error", "description": "Detected too many database errors. It's recommended to restart service.", "fields": [{"name": "Host", "value": app.config.domain, "inline": True}, {"name": "Unique ID", "value": app.config.abbr, "inline": True}, {"name": "Version", "value": app.version, "inline": True}], "color": int(app.config.hex_color, 16), "timestamp": datetime.now(timezone.utc).isoformat()}]}), {"Content-Type": "application/json"}, None)
 
                 if len(app.state.dberr) % 100 == 0:
                     app.state.dberr = []
@@ -147,12 +147,12 @@ async def tracebackHandler(request: Request, exc: Exception, err: str):
             return JSONResponse({"error": "Service Unavailable"}, status_code = 503)
 
         else:
-            logger.error(f"[{app.config.abbr}] {err_hash} [{str(datetime.now())}]\nRequest IP: {request.client.host}\nRequest URL: {str(request.url)}\n{err}")
+            logger.error(f"[{app.config.abbr}] {err_hash} [{datetime.now(timezone.utc).isoformat()}]\nRequest IP: {request.client.host}\nRequest URL: {str(request.url)}\n{err}")
 
             if app.redis.lpos("session_errs", err_hash) is None:
                 app.redis.lpush("session_errs", err_hash)
                 if app.config.webhook_error != "":
-                    opqueue.queue(app, "post", app.config.webhook_error, app.config.webhook_error, json.dumps({"embeds": [{"title": "Runtime Error", "description": f"```{err}```", "fields": [{"name": "Host", "value": app.config.domain, "inline": True}, {"name": "Unique ID", "value": app.config.abbr, "inline": True}, {"name": "Version", "value": app.version, "inline": True}, {"name": "Request IP", "value": f"`{request.client.host}`", "inline": False}, {"name": "Request URL", "value": str(request.url), "inline": False}], "footer": {"text": err_hash}, "color": int(app.config.hex_color, 16), "timestamp": str(datetime.now())}]}), {"Content-Type": "application/json"}, None)
+                    opqueue.queue(app, "post", app.config.webhook_error, app.config.webhook_error, json.dumps({"embeds": [{"title": "Runtime Error", "description": f"```{err}```", "fields": [{"name": "Host", "value": app.config.domain, "inline": True}, {"name": "Unique ID", "value": app.config.abbr, "inline": True}, {"name": "Version", "value": app.version, "inline": True}, {"name": "Request IP", "value": f"`{request.client.host}`", "inline": False}, {"name": "Request URL", "value": str(request.url), "inline": False}], "footer": {"text": err_hash}, "color": int(app.config.hex_color, 16), "timestamp": datetime.now(timezone.utc).isoformat()}]}), {"Content-Type": "application/json"}, None)
 
             return JSONResponse({"error": "Internal Server Error"}, status_code = 500)
     except:
@@ -258,7 +258,7 @@ class HubMiddleware(BaseHTTPMiddleware):
                 if response_counter >= 20 and avg_response_time > 0.5 and app.redis.get("avgrt:alerted") is None:
                     app.redis.set("avgrt:alerted", 1)
                     app.redis.expire("avgrt:alerted", 1800)
-                    opqueue.queue(app, "post", app.config.webhook_error, app.config.webhook_error, json.dumps({"embeds": [{"title": "Degraded Performance", "description": f"Degraded performance detected. It's recommended to restart service.\n\nAverage response time: {int(avg_response_time * 1000)}ms (last 30 minutes)", "fields": [{"name": "Host", "value": app.config.domain, "inline": True}, {"name": "Unique ID", "value": app.config.abbr, "inline": True}, {"name": "Version", "value": app.version, "inline": True}], "color": int(app.config.hex_color, 16), "timestamp": str(datetime.now())}]}), {"Content-Type": "application/json"}, None)
+                    opqueue.queue(app, "post", app.config.webhook_error, app.config.webhook_error, json.dumps({"embeds": [{"title": "Degraded Performance", "description": f"Degraded performance detected. It's recommended to restart service.\n\nAverage response time: {int(avg_response_time * 1000)}ms (last 30 minutes)", "fields": [{"name": "Host", "value": app.config.domain, "inline": True}, {"name": "Unique ID", "value": app.config.abbr, "inline": True}, {"name": "Version", "value": app.version, "inline": True}], "color": int(app.config.hex_color, 16), "timestamp": datetime.now(timezone.utc).isoformat()}]}), {"Content-Type": "application/json"}, None)
 
             for middleware in app.external_middleware["response_ok"]:
                 try: