+++ src/apis/tracker/trucky.py
@@ -8,7 +8,7 @@ import json
 import random
 import time
 import traceback
-from datetime import datetime
+from datetime import datetime, timezone
 from random import randint
 from typing import Optional
 from urllib.parse import parse_qs
@@ -471,7 +471,7 @@ async def handle_new_job(request, response, original_data, data, bypass_tracker_
                                         {"name": ml.ctr(request, "net_profit"), "value": f"{munit}{tseparator(int(revenue))}", "inline": True},
                                         {"name": ml.ctr(request, "time_spent"), "value": original_data["data"]["duration"], "inline": True}],
                                     "footer": {"text": multiplayer}, "color": int(app.config.hex_color, 16),\
-                                    "timestamp": str(datetime.now()), "image": {"url": imgurl}}]}
+                                    "timestamp": datetime.now(timezone.utc).isoformat(), "image": {"url": imgurl}}]}
                     elif app.config.distance_unit == "metric":
                         data = {"embeds": [{"title": f"{ml.ctr(request, 'delivery')} #{logid}",
                                 "url": dlglink,
@@ -485,7 +485,7 @@ async def handle_new_job(request, response, original_data, data, bypass_tracker_
                                         {"name": ml.ctr(request, "net_profit"), "value": f"{munit}{tseparator(int(revenue))}", "inline": True},
                                         {"name": ml.ctr(request, "time_spent"), "value": original_data["data"]["duration"], "inline": True}],
                                     "footer": {"text": multiplayer}, "color": int(app.config.hex_color, 16),\
-                                    "timestamp": str(datetime.now()), "image": {"url": imgurl}}]}
+                                    "timestamp": datetime.now(timezone.utc).isoformat(), "image": {"url": imgurl}}]}
                     try:
                         if app.config.hook_delivery_log.channel_id != "":
                             durl = f"https://discord.com/api/v10/channels/{app.config.hook_delivery_log.channel_id}/messages"
@@ -513,7 +513,7 @@ async def handle_new_job(request, response, original_data, data, bypass_tracker_
                                         {"name": ml.tr(request, "net_profit", force_lang = language), "value": f"{munit}{tseparator(int(revenue))}", "inline": True},
                                         {"name": ml.tr(request, "time_spent", force_lang = language), "value": original_data["data"]["duration"], "inline": True}],
                                     "footer": {"text": umultiplayer}, "color": int(app.config.hex_color, 16),\
-                                    "timestamp": str(datetime.now()), "image": {"url": imgurl}}]}
+                                    "timestamp": datetime.now(timezone.utc).isoformat(), "image": {"url": imgurl}}]}
                     elif app.config.distance_unit == "metric":
                         data = {"embeds": [{"title": f"{ml.tr(request, 'delivery', force_lang = language)} #{logid}",
                                 "url": dlglink,
@@ -527,7 +527,7 @@ async def handle_new_job(request, response, original_data, data, bypass_tracker_
                                         {"name": ml.tr(request, "net_profit", force_lang = language), "value": f"{munit}{tseparator(int(revenue))}", "inline": True},
                                         {"name": ml.tr(request, "time_spent", force_lang = language), "value": original_data["data"]["duration"], "inline": True}],
                                     "footer": {"text": umultiplayer}, "color": int(app.config.hex_color, 16),\
-                                    "timestamp": str(datetime.now()), "image": {"url": imgurl}}]}
+                                    "timestamp": datetime.now(timezone.utc).isoformat(), "image": {"url": imgurl}}]}
                     if await CheckNotificationEnabled(request, "dlog", uid):
                         await SendDiscordNotification(request, uid, data)
                     await UpdateRoleConnection(request, discordid)