+++ src/apis/user/notification.py
@@ -4,6 +4,7 @@
 import copy
 import json
 import math
+from datetime import datetime, timezone
 from typing import Optional
 
 from fastapi import Header, Request, Response
@@ -291,7 +292,7 @@ async def post_settings_enable(request: Request, response: Response, notificatio
                 r = await arequests.post(app, f"https://discord.com/api/v10/channels/{channelid}/messages", headers = headers, data=json.dumps({"embeds": [{"title": ml.tr(request, "notification", force_lang = await GetUserLanguage(request, uid)),
                 "description": ml.tr(request, "discord_notification_enabled", force_lang = await GetUserLanguage(request, uid)), \
                 "footer": {"text": app.config.name, "icon_url": app.config.logo_url}, \
-                "timestamp": str(datetime.now()), "color": int(app.config.hex_color, 16)}]}))
+                "timestamp": datetime.now(timezone.utc).isoformat(), "color": int(app.config.hex_color, 16)}]}))
             except:
                 pass